+++ Commands/ClientSidePages/ConvertToClientSidePage.cs
@@ -11,6 +11,7 @@ using Microsoft.SharePoint.Client;
 using System.Xml.Serialization;
 using SharePointPnP.Modernization.Framework;
 using System.IO;
+using SharePointPnP.Modernization.Framework.Telemetry.Observers;
 
 namespace SharePointPnP.PowerShell.Commands.ClientSidePages
 {
@@ -28,16 +29,28 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
     SortOrder = 2)]
     [CmdletExample(
     Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -Overwrite -AddPageAcceptBanner",
-    Remarks = "Converts a wiki page named 'somepage' to a client side page and adds the page accept banner web part on top of the page. This requires that the SPFX solution holding the web part (https://github.com/SharePoint/sp-dev-modernization/blob/master/Solutions/PageTransformationUI/assets/sharepointpnp-pagetransformation-client.sppkg?raw=true) has been installed to the tenant app catalog.",
+    Remarks = "Converts a wiki page named 'somepage' to a client side page and adds the page accept banner web part on top of the page. This requires that the SPFX solution holding the web part (https://github.com/SharePoint/sp-dev-modernization/blob/master/Solutions/PageTransformationUI/assets/sharepointpnp-pagetransformation-client.sppkg?raw=true) has been installed to the tenant app catalog",
     SortOrder = 3)]
     [CmdletExample(
     Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -Overwrite -CopyPageMetadata",
     Remarks = "Converts a wiki page named 'somepage' to a client side page, including the copying of the page metadata (if any)",
     SortOrder = 4)]
     [CmdletExample(
+    Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -Library ""SiteAssets"" -Folder ""Folder1"" -Overwrite",
+    Remarks = "Converts a web part page named 'somepage' living inside the SiteAssets library in a folder named folder1 into a client side page",
+    SortOrder = 5)]
+    [CmdletExample(
     Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -Overwrite -TargetWebUrl https://contoso.sharepoint.com/sites/targetmodernsite",
     Remarks = "Converts a wiki page named 'somepage' to a client side page in the https://contoso.sharepoint.com/sites/targetmodernsite site",
-    SortOrder = 5)]
+    SortOrder = 6)]
+    [CmdletExample(
+    Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -LogType File -LogFolder c:\temp -LogVerbose -Overwrite",
+    Remarks = "Converts a web part page named 'somepage' and creates a log file in c:\temp using verbose logging",
+    SortOrder = 7)]
+    [CmdletExample(
+    Code = @"PS:> ConvertTo-PnPClientSidePage -Identity ""somepage.aspx"" -LogType SharePoint -LogSkipFlush",
+    Remarks = "Converts a web part page named 'somepage' and creates a log file in SharePoint but skip the actual write. Use this option to make multiple ConvertTo-PnPClientSidePage invocations create a single log",
+    SortOrder = 8)]
     public class ConvertToClientSidePage : PnPWebCmdlet
     {
         private Assembly modernizationAssembly;
@@ -46,6 +59,12 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
 
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0, HelpMessage = "The name of the page to convert")]
         public PagePipeBind Identity;
+
+        [Parameter(Mandatory = false, ValueFromPipeline = true, Position = 0, HelpMessage = "The name of the library containing the page. If SitePages then please omit this parameter")]
+        public string Library;
+
+        [Parameter(Mandatory = false, ValueFromPipeline = true, Position = 0, HelpMessage = "The folder to load the provided page from. If not provided all folders are searched")]
+        public string Folder;
 
         [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true, HelpMessage = "Path and name of the web part mapping file driving the transformation", ParameterSetName = "WebPartMappingFile")]
         public string WebPartMappingFile;
@@ -62,23 +81,35 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
         [Parameter(Mandatory = false, HelpMessage = "Adds the page accept banner web part. The actual web part is specified in webpartmapping.xml file")]
         public SwitchParameter AddPageAcceptBanner = false;
 
-        [Parameter(Mandatory = false, HelpMessage = "By default the item level permissions on a page are copied to the created client side page. Use this switch to prevent the copy.")]
+        [Parameter(Mandatory = false, HelpMessage = "By default the item level permissions on a page are copied to the created client side page. Use this switch to prevent the copy")]
         public SwitchParameter SkipItemLevelPermissionCopyToClientSidePage = false;
 
-        [Parameter(Mandatory = false, HelpMessage = "Clears the cache. Can be needed if you've installed a new web part to the site and want to use that in a custom webpartmapping file. Restarting your PS session has the same effect.")]
+        [Parameter(Mandatory = false, HelpMessage = "Clears the cache. Can be needed if you've installed a new web part to the site and want to use that in a custom webpartmapping file. Restarting your PS session has the same effect")]
         public SwitchParameter ClearCache = false;
 
-        [Parameter(Mandatory = false, HelpMessage = "Copies the page metadata to the created modern page.")]
+        [Parameter(Mandatory = false, HelpMessage = "Copies the page metadata to the created modern page")]
         public SwitchParameter CopyPageMetadata = false;
 
-        [Parameter(Mandatory = false, HelpMessage = "Uses the community script editor (https://github.com/SharePoint/sp-dev-fx-webparts/tree/master/samples/react-script-editor) as replacement for the classic script editor web part.")]
+        [Parameter(Mandatory = false, HelpMessage = "Uses the community script editor (https://github.com/SharePoint/sp-dev-fx-webparts/tree/master/samples/react-script-editor) as replacement for the classic script editor web part")]
         public SwitchParameter UseCommunityScriptEditor = false;
 
-        [Parameter(Mandatory = false, HelpMessage = "By default summarylinks web parts are replaced by QuickLinks, but you can transform to plain html by setting this switch.")]
+        [Parameter(Mandatory = false, HelpMessage = "By default summarylinks web parts are replaced by QuickLinks, but you can transform to plain html by setting this switch")]
         public SwitchParameter SummaryLinksToHtml = false;
 
-        [Parameter(Mandatory = false, HelpMessage = "Url of the target web that will receive the modern page. Defaults to null which means in-place transformation.")]
+        [Parameter(Mandatory = false, HelpMessage = "Url of the target web that will receive the modern page. Defaults to null which means in-place transformation")]
         public string TargetWebUrl;
+
+        [Parameter(Mandatory = false, HelpMessage = "Allows to generate a transformation log (File | SharePoint)")]
+        public ClientSidePageTransformatorLogType LogType = ClientSidePageTransformatorLogType.None;
+
+        [Parameter(Mandatory = false, HelpMessage = "Folder in where the log file will be created (if LogType==File)")]
+        public string LogFolder = "";
+
+        [Parameter(Mandatory = false, HelpMessage = "By default each cmdlet invocation will result in a log file, use the -SkipLogFlush to delay the log flushing. The first call without -SkipLogFlush will then write all log entries to a single log")]
+        public SwitchParameter LogSkipFlush = false;
+
+        [Parameter(Mandatory = false, HelpMessage = "Configure logging to include verbose log entries")]
+        public SwitchParameter LogVerbose = false;
 
         protected override void ExecuteCmdlet()
         {
@@ -86,6 +117,8 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
             FixAssemblyResolving();
 
             // Load the page to transform
+            Identity.Library = this.Library;
+            Identity.Folder = this.Folder;
             var page = Identity.GetPage(this.ClientContext.Web);
 
             if (page == null)
@@ -140,6 +173,16 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
                 pageTransformator = new PageTransformator(this.ClientContext, targetContext, webPartMappingModel);
             }
 
+            // Setup logging
+            if (this.LogType == ClientSidePageTransformatorLogType.File)
+            {
+                pageTransformator.RegisterObserver(new MarkdownObserver(folder:this.LogFolder, includeDebugEntries:this.LogVerbose));
+            }
+            else if (this.LogType == ClientSidePageTransformatorLogType.SharePoint)
+            {
+                pageTransformator.RegisterObserver(new MarkdownToSharePointObserver(targetContext ?? this.ClientContext, includeDebugEntries: this.LogVerbose));
+            }
+
             // Setup Transformation information
             PageTransformationInformation pti = new PageTransformationInformation(page)
             {
@@ -165,6 +208,15 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
             }
 
             string serverRelativeClientPageUrl = pageTransformator.Transform(pti);
+
+            // Flush log
+            if (this.LogType != ClientSidePageTransformatorLogType.None)
+            {
+                if (!this.LogSkipFlush)
+                {
+                    pageTransformator.FlushObservers();
+                }
+            }
 
             // Output the server relative url to the newly created page
             if (!string.IsNullOrEmpty(serverRelativeClientPageUrl))